Flight Summary from Airports with the Most Cancellations

This dataset contains information on the minimum and maximum number of flights that originated from the top 10 U.S. airports with the highest number of flight cancellations. The airports included are Denver (DEN), Orlando (MCO), Seattle (SEA), Atlanta (ATL), Dallas/Fort Worth (DFW), Chicago O'Hare (ORD), Las Vegas (LAS), Los Angeles (LAX), and Phoenix (PHX).

Usage

flights

Format

`flights` A data frame with 120 rows and 6 columns:

origin

The origin airport for that flight

month

The month of the flight

long

Longitude of the airport

lat

Latitude of the airport

min_flights

The minimum number of flights that originated from the airport

max_flights

The maximum number of flights that originated from the airport