Monthly Train Station Patronage 2023-2024

This dataset provides a comprehensive monthly summary of daily patronage at each train station in Victoria for the fiscal year 2023-2024. The number of patronage is definded by the total number of boardings and alightings at each station. This dataset includes detailed breakdowns by day types such as weekdays, weekends, and different peak periods, enhancing transit management and planning capabilities. Note the influence of station closures which may skew or omit data on particular days.

Usage

data(train)

Format

A data frame with each row representing aggregated monthly data per station, containing:

month

Operating month

year

Financial year

station_name

Name of the train station.

lat

Latitude of the train station.

long

Longitude of the train station.

services

Number of unique services passing through the station.

mode

Transportation mode, such as 'Metro', 'VLine', or both.

min_monthly

Minimum daily patronage by month.

max_monthly

Maximum daily patronage by month.

min_weekday

Minimum daily patronage on a typical weekday.

max_weekday

Maximum daily patronage on a typical weekday.

min_weekend

Minimum daily patronage on weekends.

max_weekend

Maximum daily patronage on weekends.

min_preAM_peak

Minimum daily patronage during pre-AM peak hours.

max_preAM_peak

Maximum daily patronage during pre-AM peak hours.

min_AM_peak

Minimum daily patronage during AM peak hours.

max_AM_peak

Maximum daily patronage during AM peak hours.

min_interpeak

Minimum daily patronage during interpeak hours.

max_interpeak

Maximum daily patronage during interpeak hours.

min_PM_peak

Minimum daily patronage during PM peak hours.

max_PM_peak

Maximum daily patronage during PM peak hours.

min_PM_late

Minimum daily patronage during late PM hours.

max_PM_late

Maximum daily patronage during late PM hours.

Source

https://discover.data.vic.gov.au/dataset/train-service-passenger-counts

Note

Stations with incomplete entries due to closure are not included in this dataset.